split pricing analysis and plots
