+++ config/config.yaml
@@ -4,10 +4,12 @@
 
 # docs in https://pypsa-eur.readthedocs.io/en/latest/configuration.html#run
 run:
-  prefix: 20250307-Pricing
+  prefix: 20250311-Pricing-Fabian
 
   name:
-  - KN2045_Bal_v4_voll
+  # - KN2045_Bal_v4
+  - KN2045_Bal_v4_24H
+  # - KN2045_Bal_v4_voll
   # - KN2045_Bal_v4_elastic5GW
   # - KN2045_Bal_v4_elasticAll
   # - KN2045_Bal_v4_elasticPWL
@@ -713,20 +715,17 @@ demand_modelling:
   voll: false
   elastic_capa: false
   elastic_all: false
-  elastic_pwl: false # false, default, lower, higher
+  elastic_pwl: default # false, default, lower, higher
   elastic_intercept: 2000 # EUR/tMWh	Price at which load vanishes
   elastic_load: 5000 # MW
   voll_price: 2000 # €/MWh
   elastic_pwl_params:
     default:
-      intercept: [8000, 400, 200]
-      slope: [80, 40, 20]
-      nominal: [52000,3000,5500]
+      intercept: [9000,500, 250]
+      slope: [100, 50, 25]
     lower:
       intercept: [16000, 800, 400]
       slope: [160, 80, 40]
-      nominal: [95, 5, 10]
     higher:
       intercept: [4000, 200, 100]
       slope: [40, 20, 10]
-      nominal: [95, 5, 10]
